+++ components/index.js
@@ -1,2 +1,3 @@
 export { default as ButtonControl } from './ButtonControl';
 export { default as ScheduleCard } from './ScheduleCard';
+export { default as NotificationsComp } from './Notifications';
